Correct casting on Card instancing

+++ src/main/java/it/polimi/ingsw/model/card/InfluenceCharacterCard.java
@@ -13,8 +13,8 @@ public class InfluenceCharacterCard extends CharacterCard {
     public InfluenceCharacterCard(CharacterCard card) {
         super(card);
         Map<String, Object> params = this.getParams();
-        this.towerInfluence = (int) params.get("tower_influence");
-        this.extraInfluence = (int) params.get("extra_influence");
+        this.towerInfluence = ((Double) params.get("tower_influence")).intValue();
+        this.extraInfluence = ((Double) params.get("extra_influence")).intValue();
         this.colorWithoutInfluence = (boolean) params.get("color_without_influence");
     }
